refactor(design-system): move token stories to tokens/stories subfolder
